requirements.txt
Django==5.0.6
djangorestframework==3.15.1
drf-yasg==1.21.7
pillow==10.3.0
djangorestframework-simplejwt==5.3.1
 setuptools==69.5.1